% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manhattanplot_backend.R
\name{make_manhattanplot_backend}
\alias{make_manhattanplot_backend}
\title{Make a Manhattan Plot}
\usage{
make_manhattanplot_backend(
  data,
  plot_name = "manhattan_plot.png",
  color1 = "cornflowerblue",
  color2 = "blue4",
  threshold_linecolor = "black"
)
}
\arguments{
\item{data}{an data frame object loaded from load_data function}

\item{plot_name}{name of the Manhattan plot file}

\item{color1}{Color 1 of the dots in the Manhattan plot}

\item{color2}{Color 2 of the dots in the Manhattan plot}

\item{threshold_linecolor}{the color of the threshold lines at 1e-5 and 1e-8}
}
\description{
This function makes a Manhattan plot with the default data frame of GWAS
summary statistics from load_data function
}
